requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
flask-openapi3==2.5.2
Flask-SQLAlchemy==3.0.5
pydantic==1.10.2
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
flask-restx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
arrow==1.2.3